create session manager to handle network session
