+++ app/src/main/java/io/locative/app/GeofancyApplication.java
@@ -2,17 +2,10 @@ package io.locative.app;
 
 import android.app.Application;
 import android.content.Context;
-import android.content.SharedPreferences;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 
 import dagger.ObjectGraph;
-import io.locative.app.network.GeofancyNetworkingWrapper;
 
 public class GeofancyApplication extends Application {
-
-    private static final String APP_PREFS = "geofancy";
-    private static final String PREF_SESSION_ID = "sessionId";
 
     private ObjectGraph mObjectGraph;
 
@@ -35,27 +28,4 @@ public class GeofancyApplication extends Application {
         ((GeofancyApplication) context.getApplicationContext()).mObjectGraph.inject(target);
     }
 
-
-    public boolean hasSession() {
-        return getPrefs().contains(PREF_SESSION_ID);
-    }
-
-    @Nullable
-    public String getSessionId() {
-        return getPrefs().getString(PREF_SESSION_ID, null);
-    }
-
-    public void setSessionId(@NonNull String sessionId) {
-        // async
-        getPrefs().edit().putString(PREF_SESSION_ID, sessionId).apply();
-    }
-
-    public void clearSession() {
-        // async
-        getPrefs().edit().remove(PREF_SESSION_ID).apply();
-    }
-
-    private SharedPreferences getPrefs() {
-        return getSharedPreferences(APP_PREFS, MODE_PRIVATE);
-    }
 }
